requirements.txt
aniso8601==10.0.1
black==25.1.0
blinker==1.9.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
filelock==3.18.0
flake8==7.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
identify==2.6.12
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
virtualenv==20.31.2
Werkzeug==3.1.3
